ref: Move classes to the colibri package.
