+++ src/main/java/org/jitsi/jicofo/conference/colibri/AbstractChannelAllocator.java
@@ -15,10 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.jicofo.conference;
+package org.jitsi.jicofo.conference.colibri;
 
 import org.jitsi.jicofo.*;
-import org.jitsi.jicofo.conference.colibri.*;
+import org.jitsi.jicofo.conference.*;
 import org.jitsi.xmpp.extensions.colibri.*;
 import org.jitsi.xmpp.extensions.jingle.*;
 import org.jitsi.utils.logging2.*;
