+++ src/main/kotlin/org/jitsi/jicofo/conference/colibri/ColibriSessionManager.kt
@@ -1,27 +1,11 @@
-/*
- * Jicofo, the Jitsi Conference Focus.
- *
- * Copyright @ 2021-Present 8x8, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jitsi.jicofo.conference
+package org.jitsi.jicofo.conference.colibri
 
 import org.jitsi.jicofo.JicofoServices
 import org.jitsi.jicofo.OctoConfig
-import org.jitsi.jicofo.TaskPools.Companion.ioPool
+import org.jitsi.jicofo.TaskPools
 import org.jitsi.jicofo.bridge.Bridge
+import org.jitsi.jicofo.conference.JitsiMeetConferenceImpl
+import org.jitsi.jicofo.conference.Participant
 import org.jitsi.jicofo.conference.source.ConferenceSourceMap
 import org.jitsi.utils.event.SyncEventEmitter
 import org.jitsi.utils.logging2.Logger
@@ -125,7 +109,7 @@ class ColibriSessionManager(
         )
 
         participant.setChannelAllocator(channelAllocator)
-        ioPool.execute(channelAllocator)
+        TaskPools.ioPool.execute(channelAllocator)
 
         if (reInvite) {
             addSourcesToOcto(participant.sources, bridgeSession)
@@ -361,6 +345,3 @@ class ColibriSessionManager(
         fun failedBridgesRemoved(count: Int)
     }
 }
-
-/** Bridge selection failed, i.e. there were no bridges available. */
-class BridgeSelectionFailedException : Exception()
